Move out MergifyPull v1 stuff to v1 engine
Many methods and attributes are used only by v1 engine.
This change moves them there.
